attacker/requirements.txt
dnspython==2.2.0
cryptography==36.0.1
bitstring==3.1.9